Refactor modal into separate file
